refactor creating subject.
