+++ dev/com.ibm.ws.security.javaeesec.cdi/src/com/ibm/ws/security/javaeesec/cdi/beans/CustomFormAuthenticationMechanism.java
@@ -10,44 +10,24 @@
  *******************************************************************************/
 package com.ibm.ws.security.javaeesec.cdi.beans;
 
-import java.security.AccessController;
-import java.security.PrivilegedAction;
-import java.util.ArrayList;
-import java.util.Hashtable;
-import java.util.Map;
-import java.util.Set;
-
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Default;
 import javax.enterprise.inject.Instance;
 import javax.enterprise.inject.spi.CDI;
 import javax.security.auth.Subject;
-import javax.security.auth.callback.Callback;
-import javax.security.auth.callback.CallbackHandler;
-import javax.security.auth.message.callback.PasswordValidationCallback;
 import javax.security.enterprise.AuthenticationException;
 import javax.security.enterprise.AuthenticationStatus;
 import javax.security.enterprise.authentication.mechanism.http.AuthenticationParameters;
 import javax.security.enterprise.authentication.mechanism.http.HttpAuthenticationMechanism;
 import javax.security.enterprise.authentication.mechanism.http.HttpMessageContext;
 import javax.security.enterprise.authentication.mechanism.http.LoginToContinue;
-import javax.security.enterprise.credential.BasicAuthenticationCredential;
 import javax.security.enterprise.credential.Credential;
-import javax.security.enterprise.credential.UsernamePasswordCredential;
-import javax.security.enterprise.identitystore.CredentialValidationResult;
-import javax.security.enterprise.identitystore.IdentityStoreHandler;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import com.ibm.websphere.ras.Tr;
 import com.ibm.websphere.ras.TraceComponent;
 import com.ibm.websphere.ras.annotation.Sensitive;
-import com.ibm.ws.security.authentication.AuthenticationConstants;
-import com.ibm.wsspi.security.token.AttributeNameConstants;
-
-// TODO: investigate whether HttpMessageContext.isAuthenticationRequest() needs to be implemented.
-// The current code assumes that all of the call to validateRequest with credential is made programmatically,
-// thus there is no consideration for isAuthenticationRequest, but need to make sure whether that's the case.
 
 @Default
 @ApplicationScoped
@@ -64,8 +44,6 @@ public class CustomFormAuthenticationMechanism implements HttpAuthenticationMech
 
         Subject clientSubject = httpMessageContext.getClientSubject();
         @SuppressWarnings("unchecked")
-        Map<String, String> msgMap = httpMessageContext.getMessageInfo().getMap();
-        CallbackHandler handler = httpMessageContext.getHandler();
         HttpServletRequest req = httpMessageContext.getRequest();
         HttpServletResponse rsp = httpMessageContext.getResponse();
         AuthenticationParameters authParams = httpMessageContext.getAuthParameters();
@@ -92,7 +70,7 @@ public class CustomFormAuthenticationMechanism implements HttpAuthenticationMech
                     status = AuthenticationStatus.SEND_CONTINUE;
                 }
             } else {
-                status = handleFormLogin(cred, rsp, msgMap, clientSubject, handler);
+                status = handleFormLogin(cred, rsp, clientSubject, httpMessageContext);
             }
         }
         return status;
@@ -112,161 +90,19 @@ public class CustomFormAuthenticationMechanism implements HttpAuthenticationMech
 
     }
 
-    private AuthenticationStatus handleFormLogin(@Sensitive Credential cred, HttpServletResponse rsp, Map<String, String> msgMap, Subject clientSubject,
-                                                 CallbackHandler handler) throws AuthenticationException {
+    private AuthenticationStatus handleFormLogin(@Sensitive Credential credential, HttpServletResponse rsp, Subject clientSubject,
+                                                 HttpMessageContext httpMessageContext) throws AuthenticationException {
         AuthenticationStatus status = AuthenticationStatus.SEND_FAILURE;
         int rspStatus = HttpServletResponse.SC_FORBIDDEN;
-        status = validateUserAndPassword(clientSubject, cred, handler);
+        status = Utils.getInstance().validateCredential(getCDI(), "defaultRealm", clientSubject, credential, httpMessageContext);
         if (status == AuthenticationStatus.SUCCESS) {
-            msgMap.put("javax.servlet.http.authType", "JASPI_AUTH");
+            httpMessageContext.getMessageInfo().getMap().put("javax.servlet.http.authType", "JASPI_AUTH");
             rspStatus = HttpServletResponse.SC_OK;
         } else {
             // TODO: Audit invalid user or password
         }
         rsp.setStatus(rspStatus);
         return status;
-    }
-
-    private AuthenticationStatus validateUserAndPassword(Subject clientSubject, @Sensitive Credential credential,
-                                                         CallbackHandler handler) throws AuthenticationException {
-        AuthenticationStatus status = AuthenticationStatus.SEND_FAILURE;
-        IdentityStoreHandler identityStoreHandler = getIdentityStoreHandler();
-        if (identityStoreHandler != null) {
-            status = validateWithIdentityStore(clientSubject, credential, identityStoreHandler, handler);
-        } else {
-            Tr.warning(tc, "JAVAEESEC_CDI_WARNING_NO_IDENTITY_STORE_HANDLER");
-        }
-        if (identityStoreHandler == null || status == AuthenticationStatus.NOT_DONE) {
-            // If an identity store is not available, fall back to the original user registry.
-            status = validateWithUserRegistry(clientSubject, credential, handler);
-        }
-        return status;
-    }
-
-    private AuthenticationStatus validateWithUserRegistry(Subject clientSubject, @Sensitive Credential credential,
-                                                          CallbackHandler handler) throws AuthenticationException {
-        AuthenticationStatus status = AuthenticationStatus.SEND_FAILURE;
-        if (handler != null) {
-            if (isSupportedCredential(credential)) {
-                PasswordValidationCallback pwcb = new PasswordValidationCallback(clientSubject, ((UsernamePasswordCredential) credential).getCaller(), ((UsernamePasswordCredential) credential).getPassword().getValue());
-                try {
-                    handler.handle(new Callback[] { pwcb });
-                    boolean isValidPassword = pwcb.getResult();
-                    if (isValidPassword) {
-                        status = AuthenticationStatus.SUCCESS;
-                    }
-                } catch (Exception e) {
-                    throw new AuthenticationException(e.toString());
-                }
-            } else {
-                // This is an error condition.
-                Tr.error(tc, "JAVAEESEC_CDI_ERROR_UNSUPPORTED_CRED", credential.getClass().getName());
-                String msg = Tr.formatMessage(tc, "JAVAEESEC_CDI_ERROR_UNSUPPORTED_CRED", credential.getClass().getName());
-                throw new AuthenticationException(msg);
-            }
-        }
-        return status;
-    }
-
-    private AuthenticationStatus validateWithIdentityStore(Subject clientSubject, @Sensitive Credential credential, IdentityStoreHandler identityStoreHandler,
-                                                           CallbackHandler handler) throws AuthenticationException {
-        AuthenticationStatus status = AuthenticationStatus.SEND_FAILURE;
-        CredentialValidationResult result = identityStoreHandler.validate(credential);
-        if (result.getStatus() == CredentialValidationResult.Status.VALID) {
-            createLoginHashMap(clientSubject, result);
-            status = AuthenticationStatus.SUCCESS;
-        } else if (result.getStatus() == CredentialValidationResult.Status.NOT_VALIDATED) {
-            // TODO: error message. no identitystore.
-            status = AuthenticationStatus.NOT_DONE;
-        } else {
-            // TODO: Audit invalid user or password
-        }
-        return status;
-    }
-
-    protected void createLoginHashMap(Subject clientSubject, CredentialValidationResult result) throws AuthenticationException {
-        Utils.validateResult(result);
-        Hashtable<String, Object> credData = getSubjectCustomData(clientSubject);
-        Set<String> groups = result.getCallerGroups();
-        String realm = result.getIdentityStoreId();
-        if (realm == null) {
-            realm = "default";
-        }
-        credData.put(AttributeNameConstants.WSCREDENTIAL_REALM, realm);
-        credData.put(AttributeNameConstants.WSCREDENTIAL_USERID, result.getCallerPrincipal().getName());
-        credData.put(AttributeNameConstants.WSCREDENTIAL_SECURITYNAME, result.getCallerUniqueId());
-        credData.put(AttributeNameConstants.WSCREDENTIAL_UNIQUEID, "user:" + realm + "/" + result.getCallerUniqueId());
-
-        credData.put(AuthenticationConstants.INTERNAL_ASSERTION_KEY, Boolean.TRUE);
-        if (groups != null && !groups.isEmpty()) {
-            if (tc.isDebugEnabled()) {
-                Tr.debug(tc, "Adding groups found in an identitystore", groups);
-            }
-            credData.put(AttributeNameConstants.WSCREDENTIAL_GROUPS, new ArrayList<String>(groups));
-        } else {
-            if (tc.isDebugEnabled()) {
-                Tr.debug(tc, "No group  found in an identitystore");
-            }
-            credData.put(AttributeNameConstants.WSCREDENTIAL_GROUPS, new ArrayList<String>());
-        }
-        return;
-    }
-
-    protected Hashtable<String, Object> getSubjectCustomData(final Subject clientSubject) {
-        Hashtable<String, Object> cred = getCustomCredentials(clientSubject);
-        if (cred == null) {
-            PrivilegedAction<Hashtable<String, Object>> action = new PrivilegedAction<Hashtable<String, Object>>() {
-
-                @Override
-                public Hashtable<String, Object> run() {
-                    Hashtable<String, Object> newCred = new Hashtable<String, Object>();
-                    clientSubject.getPrivateCredentials().add(newCred);
-                    return newCred;
-                }
-            };
-            cred = AccessController.doPrivileged(action);
-        }
-        return cred;
-    }
-
-    protected Hashtable<String, Object> getCustomCredentials(final Subject clientSubject) {
-        if (clientSubject == null)
-            return null;
-        PrivilegedAction<Hashtable<String, Object>> action = new PrivilegedAction<Hashtable<String, Object>>() {
-
-            @SuppressWarnings("unchecked")
-            @Override
-            public Hashtable<String, Object> run() {
-                Set s = clientSubject.getPrivateCredentials(Hashtable.class);
-                if (s == null || s.isEmpty()) {
-                    if (tc.isDebugEnabled())
-                        Tr.debug(tc, "Subject has no Hashtable with custom credentials, return null.");
-                    return null;
-                } else {
-                    Hashtable t = (Hashtable) s.iterator().next();
-                    return t;
-                }
-            }
-        };
-        Hashtable<String, Object> cred = AccessController.doPrivileged(action);
-        return cred;
-    }
-
-    private IdentityStoreHandler getIdentityStoreHandler() {
-        IdentityStoreHandler identityStoreHandler = null;
-        Instance<IdentityStoreHandler> storeHandlerInstance = getCDI().select(IdentityStoreHandler.class);
-        if (storeHandlerInstance.isUnsatisfied() == false && storeHandlerInstance.isAmbiguous() == false) {
-            identityStoreHandler = storeHandlerInstance.get();
-        }
-        return identityStoreHandler;
-    }
-
-    private boolean isSupportedCredential(@Sensitive Credential cred) {
-        if (cred != null && (cred instanceof UsernamePasswordCredential || cred instanceof BasicAuthenticationCredential)) {
-            return true;
-        } else {
-            return false;
-        }
     }
 
     protected CDI getCDI() {
